+++ src/app/commands/poap/POAP.ts
@@ -1,9 +1,4 @@
-import {
-	CommandContext,
-	CommandOptionType,
-	SlashCommand,
-	SlashCreator,
-} from 'slash-create';
+import { CommandContext, CommandOptionType, SlashCommand, SlashCreator } from 'slash-create';
 import ServiceUtils from '../../utils/ServiceUtils';
 import StartPOAP from '../../service/poap/StartPOAP';
 import ValidationError from '../../errors/ValidationError';
@@ -11,14 +6,13 @@ import EarlyTermination from '../../errors/EarlyTermination';
 import EndPOAP from '../../service/poap/EndPOAP';
 import DistributePOAP from '../../service/poap/DistributePOAP';
 import ConfigPOAP from '../../service/poap/ConfigPOAP';
-import discordServerIds from '../../service/constants/discordServerIds';
+import SchedulePOAP from '../../service/poap/SchedulePOAP';
 
 module.exports = class poap extends SlashCommand {
 	constructor(creator: SlashCreator) {
 		super(creator, {
 			name: 'poap',
 			description: 'Receive a list of all attendees in the specified voice channel and optionally send out POAP links',
-			guildIDs: [discordServerIds.banklessDAO, discordServerIds.discordBotGarage],
 			options: [
 				{
 					name: 'config',
@@ -60,6 +54,19 @@ module.exports = class poap extends SlashCommand {
 							type: CommandOptionType.USER,
 							description: 'The user that should have access to poap commands.',
 							required: false,
+						},
+					],
+				},
+				{
+					name: 'schedule',
+					type: CommandOptionType.SUB_COMMAND,
+					description: 'Schedule a POAP event, upload the PNG image to be minted, and get the links.txt file over email.',
+					options: [
+						{
+							name: 'mint-copies',
+							type: CommandOptionType.INTEGER,
+							description: 'The number of POAPs to be minted for all of the participants. Best to overestimate.',
+							required: true,
 						},
 					],
 				},
@@ -109,19 +116,23 @@ module.exports = class poap extends SlashCommand {
 				console.log(`/poap config ${ctx.user.username}#${ctx.user.discriminator}`);
 				authorizedRoles = [ctx.options.config['role-1'], ctx.options.config['role-2'], ctx.options.config['role-3']];
 				authorizedUsers = [ctx.options.config['user-1'], ctx.options.config['user-2'], ctx.options.config['user-3']];
-				command = ConfigPOAP(guildMember, authorizedRoles, authorizedUsers);
+				command = ConfigPOAP(ctx, guildMember, authorizedRoles, authorizedUsers);
+				break;
+			case 'schedule':
+				console.log(`/poap schedule ${ctx.user.username}#${ctx.user.discriminator}`);
+				command = SchedulePOAP(ctx, guildMember, ctx.options.schedule['mint-copies']);
 				break;
 			case 'start':
 				console.log(`/poap start ${ctx.user.username}#${ctx.user.discriminator}`);
-				command = StartPOAP(guildMember, ctx.options.start.event);
+				command = StartPOAP(ctx, guildMember, ctx.options.start.event);
 				break;
 			case 'end':
 				console.log(`/poap end ${ctx.user.username}#${ctx.user.discriminator}`);
-				command = EndPOAP(guildMember);
+				command = EndPOAP(ctx, guildMember);
 				break;
 			case 'distribute':
 				console.log(`/poap distribute ${ctx.user.username}#${ctx.user.discriminator}`);
-				command = DistributePOAP(guildMember);
+				command = DistributePOAP(ctx, guildMember);
 				break;
 			default:
 				return ctx.send(`${ctx.user.mention} Please try again.`);
@@ -139,8 +150,6 @@ module.exports = class poap extends SlashCommand {
 				return ctx.send('POAPS sent. Expect delivery shortly.');
 			} else if (result === 'POAP_END') {
 				return ctx.send('POAP event ended. POAPs will be delivered at a later time.');
-			} else {
-				return ctx.send(`${ctx.user.mention} DM sent!`);
 			}
 		}).catch(e => {
 			if (e instanceof ValidationError) {